backend/requirements.txt
fastapi
uvicorn[standard]
stellar-sdk
soroban-rpc-client
requests
python-multipart
python-dotenv 